// src/core/permissions/permission.types.ts
export type AppRole =
  | 'RESIDENT_OWNER'
  | 'RESIDENT_TENANT'
  | 'RESIDENT_FAMILY'
  | 'SECURITY_GUARD'
  | 'SECURITY_SUPERVISOR'
  | 'FACILITY_MANAGER'
  | 'COMMITTEE_MEMBER'
  | 'SECRETARY'
  | 'CHAIRPERSON'
  | 'TREASURER'
  | 'SOCIETY_ADMIN'
  | 'SUPER_ADMIN'
  | 'ELECTION_OFFICER'
  | 'AUDITOR'
  | 'VENDOR_USER'
  | 'STAFF_USER';